pyproject.toml
[project]
name = "graph-agent"
version = "0.1.0"
description = "AI-powered CLI tool for creating brand-compliant charts from data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "click>=8.1.0",
    "matplotlib>=3.8.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
graph-agent = "graph_agent.cli:main"

[tool.setuptools.packages.find]
include = ["graph_agent*"]
exclude = ["backlog*", "tests*"]
